+++ src/Component/DatePicker/index.tsx
@@ -1,13 +1,15 @@
 import React, { useState, useEffect, useRef } from 'react';
-import TextField from 'react-md/lib/TextFields/TextField';
+import { connect } from 'react-redux';
 import ReactDatePicker, { ReactDatePickerProps } from 'react-datepicker';
 import 'react-datepicker/dist/react-datepicker.css';
-import './Setup';
 import OutsideClickHandler from 'react-outside-click-handler';
+import { InjectedIntlProps, injectIntl } from 'react-intl';
+
+import './Setup';
 import International from '../International';
 import Global from 'Store/Reducer/Global';
 import { LANGUAGE, DATE_FORMAT } from 'Types/Language';
-import { connect } from 'react-redux';
+import './_styles.scss'
 
 const mapLanguageToDateFormat = (language: LANGUAGE) => {
   switch (language) {
@@ -20,20 +22,24 @@ const mapLanguageToDateFormat = (language: LANGUAGE) => {
 
 namespace DatePicker {
   export type Props = {
+    label?: string;
+    customClearButtonId?: string;
     value: Date;
-    onChange: (newDate: Date) => void;
     state?: { language: string };
-  } & ReactDatePickerProps;
+    onChange: (newDate: Date, oldDate: Date, event: (React.SyntheticEvent<any> | undefined)) => void;
+  } & ReactDatePickerProps & InjectedIntlProps;
 }
 
 const DatePicker: React.FC<DatePicker.Props> = ({
-  customInput,
-  dateFormat,
-  value,
-  onChange,
-  state: { language },
-  ...rest
-}) => {
+                                                  label,
+                                                  customClearButtonId,
+                                                  dateFormat,
+                                                  value,
+                                                  onChange,
+                                                  intl: {formatMessage},
+                                                  state: { language },
+                                                  ...rest
+                                                }) => {
   const [date, setDate] = useState<Date>(value);
   const [open, setOpen] = useState(false);
   const [id] = useState(Math.floor(Math.random() * 100));
@@ -51,6 +57,71 @@ const DatePicker: React.FC<DatePicker.Props> = ({
     }
   );
 
+  useEffect(() => {
+    document.getElementById(customClearButtonId)?.addEventListener('click', () => {
+      // The clear method does exists => its just not in the typing
+      // @ts-ignore
+      datePickerRef.current.clear();
+    });
+    const input = getInput();
+    input.addEventListener('keydown', _handleKeyupEventOnCustomInputField);
+    input.addEventListener('blur', _handleBlurEventOnCustomInputField);
+    const finalLabel = label !== null ? label : formatMessage({id: 'form.datePicker.label'});
+    getInputParent().setAttribute('data-label', finalLabel);
+  }, []);
+
+  const _handleKeyupEventOnCustomInputField = (e: KeyboardEvent) => {
+    const input = getInput();
+    if (document.activeElement.id === input.id && e.keyCode === 9 && e.shiftKey) {
+      e.preventDefault();
+      e.stopPropagation();
+    }
+  };
+
+  const _handleBlurEventOnCustomInputField = () => {
+    checkRawHasValue();
+    if (datePickerRef.current.isCalendarOpen() === false) {
+      setActive(false);
+    }
+  };
+
+  const checkRawHasValue = () => {
+    const input = getInput();
+    if (input.value === '') {
+      setHasValue(false);
+    } else {
+      setHasValue(true);
+    }
+  }
+
+  const setHasValue = (hasValue: boolean) => {
+    const inputParent = getInputParent();
+    if (hasValue) {
+      inputParent.classList.add('has-value');
+    } else {
+      inputParent.classList.remove('has-value');
+    }
+  }
+
+  const setActive = (isActive: boolean) => {
+    const inputParent = getInputParent();
+    if (isActive) {
+      inputParent.classList.add('is-active');
+    } else {
+      inputParent.classList.remove('is-active');
+    }
+  }
+
+  const getInput = () => {
+    // The clear method does exists => its just not in the typing
+    // @ts-ignore
+    return datePickerRef.current.input as HTMLInputElement;
+  };
+
+  const getInputParent = () => {
+    return getInput().parentElement;
+  };
+
   return (
     <International>
       {({ formatMessage }) => (
@@ -59,12 +130,21 @@ const DatePicker: React.FC<DatePicker.Props> = ({
             {...rest}
             ref={datePickerRef}
             selected={date}
-            onChange={(newDate) => {
-              setDate(newDate);
-              onChange(newDate);
+            onChange={(newDate, event) => {
+              setHasValue(newDate !== null);
+              setDate(newDate as Date);
+              onChange(newDate as Date, date, event);
               if (datePickerRef.current.props.shouldCloseOnSelect) {
                 setOpen(false);
               }
+            }}
+            onCalendarOpen={() => {
+              setActive(true);
+              // addEventListenerToCustomInputField();
+            }}
+            onCalendarClose={() => {
+              setActive(false);
+              // removeEventListenerToCustomInputField();
             }}
             dateFormat={dateFormat || mapLanguageToDateFormat(
               language as LANGUAGE
@@ -72,12 +152,7 @@ const DatePicker: React.FC<DatePicker.Props> = ({
             locale='de'
             open={open}
             id={id.toString()}
-            customInput={
-              customInput ||
-              <TextField
-                label={formatMessage({ id: 'date' })}
-                id={id}/>
-            }/>
+            customInput={<input className='md-full-width md-text md-text-field customDatePickerInputField'/>}/>
         </OutsideClickHandler>
       )}
     </International>
@@ -91,4 +166,4 @@ const enhance = connect(
   })
 );
 
-export default enhance(DatePicker);
+export default injectIntl(enhance(DatePicker));