+++ src/Component/Header/AppHeader.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import AppToolbar from './AppToolbar';
 
 export default ({ leftTools, rightTools }) => (
-  <div className="app-header">
+  <div className='app-header'>
     <AppToolbar
       LeftTools={leftTools}
       RightTools={rightTools}